requirements.txt
Flask==2.3.3
Flask-CORS==5.0.0
numpy>=1.23.5,<2.0.0
matplotlib~=3.10.1
tensorflow-cpu==2.12.0
fastapi>=0.100.0
uvicorn>=0.15.0
firebase-admin>=5.3.0
python-dotenv>=0.19.0
supabase>=1.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic-settings>=2.1.0
pydantic==2.5.2
pydantic[email]>=2.5.2
email-validator>=2.0.0
sendgrid>=6.9.7